[DIDA] fix. 게시글 작성 > 게시글 작성 기능 및 이동 로직 추가

+++ data/src/main/java/com/dida/data/model/detail/GetDetailNFTResponse.kt
@@ -3,7 +3,7 @@ package com.dida.data.model.detail
 import com.google.gson.annotations.SerializedName
 
 data class GetDetailNFTResponse(
-    @SerializedName("cardId") val cardId: Int,
+    @SerializedName("cardId") val cardId: Long,
     @SerializedName("contracts") val contracts: String?,
     @SerializedName("description") val description: String,
     @SerializedName("id") val id: String,
